+++ example-app/SantokuApp/src/screens/team/TeamDetailScreen.tsx
@@ -1,14 +1,27 @@
 import React from 'react';
+import {StyleSheet, View} from 'react-native';
+import {Button, Text} from 'react-native-elements';
 import {useNavigateToDemoStackScreen} from 'screens/useNavigateToScreen';
-
-import {TeamDetailTemplate} from './TeamDetailTemplate';
 
 const ScreenName = 'TeamDetail';
 const Screen: React.FC = () => {
   const onGoToDemoScreen = useNavigateToDemoStackScreen('Demo');
 
-  return <TeamDetailTemplate testID="TeamDetailScreen" onGoToDemoScreen={onGoToDemoScreen} />;
+  return (
+    <View style={styles.container} testID="TeamDetailScreen">
+      <Text h2>開発中</Text>
+      <Button title="Go to Demo" onPress={onGoToDemoScreen} />
+    </View>
+  );
 };
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+});
 
 export const TeamDetailScreen = {
   component: Screen,
